Refactor: created interface from UserService
